Add Android support for KUtrace
This commit adds support for running KUtrace on Android. Build and setup
instructions are described in linux/android/README.md. This commit also
fixes some bugs in the postprocessing such as assuming 32-bit time
counters for ARM64.

Two sets of patches are added: with and without kernel ABI changes. The
patches without ABI changes should apply cleaner to kernel versions
which don't match the ones specified in the README.

This commit also improves the base40.cc program to print delimiting
pairs of labels.